+++ src/main/java/com/janosgyerik/counters/EventRepo.java
@@ -4,9 +4,9 @@ public interface EventRepo {
   /**
    * Check if an event was already logged for the user's counter, period, action type.
    */
-  boolean exists(User user, String counterId, int periodId, ActionType... actionTypes);
+  boolean exists(User user, String counterId, long periodId, ActionType... actionTypes);
 
-  void add(User user, String counterId, int periodId, ActionType actionType, int valueBefore);
+  void add(User user, String counterId, long periodId, ActionType actionType, int valueBefore);
 
   int size();
 }
