+++ src/main/java/com/janosgyerik/counters/Period.java
@@ -3,5 +3,5 @@ package com.janosgyerik.counters;
 import java.util.Date;
 
 public interface Period {
-  int computeId(Date date, int utcOffset);
+  long computeId(Date date, int utcOffset);
 }
